requirements.txt
altgraph==0.17
APScheduler==3.6.3
asgiref==3.2.7
autopep8==1.5.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
flake8==3.8.3
future==0.18.2
idna==2.9
importlib-metadata==1.7.0
mccabe==0.6.1
pefile==2019.4.18
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pyinstaller==4.2
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
tzlocal==2.0.0
urllib3==1.25.8
zipp==3.1.0
